requirements/common.txt
Django==5.1.4
djangorestframework==3.15.2
python-decouple==3.8
psycopg2-binary==2.9.10
factory_boy==3.3.1
django-debug-toolbar==5.0.1
